+++ examples/demo7/src/main.js
@@ -1,0 +1,13 @@
+import { render, h } from 'preact'
+import App from './components/App'
+
+export default function createApp() {
+  const container = document.createElement('div')
+  container.id = 'app'
+  document.body.appendChild(container)
+
+  render(<App />, container)
+}
+
+"undefined" != typeof wx && wx.getSystemInfoSync || createApp()
+